+++ EPS Homebridge.indigoPlugin/Contents/Server Plugin/lib/eps.py
@@ -15,7 +15,7 @@ from support import support
 
 
 class eps:
-	VERSION = "2.3.0"
+	VERSION = "2.3.1"
 	
 	#
 	# Initialize the  class